Rename the exception in import
