+++ src/test/java/seedu/address/model/statistics/DailyRevenueListTest.java
@@ -13,7 +13,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import seedu.address.model.statistics.exception.DailyRevenueNotFoundException;
+import seedu.address.model.statistics.exceptions.DailyRevenueNotFoundException;
 import seedu.address.testutil.StatisticsBuilder;
 
 public class DailyRevenueListTest {
